Remove all on state load events
Removed all on state load events to avoid crashes on load state.

This commit should be undone (manually if possible to make sure everything works correclty) when the onstateload event will be synchronous with the emu thread.

For now, the expected on state load functions are called when the race frame counter isn't incrementing (or staying unchanged) on frame advance